rename SPI to common
