+++ quarkus-asyncapi-common/pom.xml
@@ -6,8 +6,7 @@
         <artifactId>quarkus-asyncapi-root</artifactId>
         <version>1.0.0-SNAPSHOT</version>
     </parent>
-    <artifactId>quarkus-asyncapi-spi</artifactId>
-    <name>Quarkus - AsyncAPI - SPI</name>
-    <description>AsyncAPI Service Provider Interface</description>
+    <artifactId>quarkus-asyncapi-common</artifactId>
+    <name>Quarkus - AsyncAPI - Common</name>
     <packaging>jar</packaging>
 </project>
